Some refactoring in the genbb::g4::primary_generator::_generate_event method
The genbb_help now supports nucleus and/or ion particle
from its generators. Here the genbb::g4::primary_generator class
(input interface of primary events in Geant4) implements
some code to support these case. This is preliminary and
seems to work for fully ionized nuclei. However, support for
partially ionized ions (example: 12C++, 6Li+...) is not
supported yet because we need to learn how to ask Geant4
to set the effective charge of the ion and not use the
charge of the nucleus (atomic number Z).

Also some preliminary code for supporting particle PDG code
is now available (however it has not been tested; consider it
as a hook for a future PDG encoding support in genbb_help).