Fix a bug and prepare event biasing
The  mctools::g4::biasing_manager  class is  added.   It  is an  empty
object for now.  It will wrap event biasing dedicated  code for Geant4
in a near future (task: François, Arnaud).

The  mctools::g4::detector_construction class  may  now instantiate  a
biasing_manager object (does nothing for now).

Testing  this  addon, I  have  discovered  a  bug  at G4  run  manager
destruction.  An exception occured while  the run manager was deleting
its internal  pointers (which is not  done in the mctools  code but in
the G4  code, so out  of our scope). The  detector_construction object
was  invoking some  invalid code  from its  destructor, with  a thrown
exception and breaking the sanity of the destruction of objects at the
end of the run.

Update some documentation files in the examples directory.