Refactor Geant4 simulation module
Refactor naming and interface in line with modern C++ standards for
clarity and purpose. Remove useless logging lines, though retain
those in multithreaded part as these partly constitute comments on
what's going on here.

Remove "erase former SD bank" flag as this performed no useful task.
The module now requires the name of the bank it will write to be unique
in the things object being processed as the module acts as a "source"
or "producer" of data. Should requirements change and an existing
bank be required, a clearer flag and simpler logic should be used.