Preliminary release of signal data support
This set of classes aims to represent analog signals from some hit detectors.
A "mctools::signal::base_signal" class is provided. It hosts the minimal
information to represent a signal hit. It inherits the "geomtools::base_hit".

Available attributes are:
- hit ID
- geom ID
- auxiliary properties
  - used to store parameters specific to the signal shape functor (see below)

Additional attributes are:
- hidden: a functor representing the shape of the analog signal
  it uses the interface if mygsl::i_unary_function
- the type ID of the function (used for factory registration)

we will evaluate this API within Falaise as the backend of
simulation a,d within the digitization process.

The "mctools::signal::base_signal" class
and the general "bank" container "mctools::signal::signal_data"
are serializable (restrictions exists so far).